fix: Markdown escape characters left in title on import
