+++ server/commands/documentImporter.ts
@@ -1,5 +1,4 @@
 import emojiRegex from "emoji-regex";
-import escapeRegExp from "lodash/escapeRegExp";
 import truncate from "lodash/truncate";
 import parseTitle from "@shared/utils/parseTitle";
 import { DocumentValidation } from "@shared/validations";
@@ -51,10 +50,7 @@ async function documentImporter({
   if (text.trim().startsWith("# ")) {
     const result = parseTitle(text);
     title = result.title;
-    text = text
-      .trim()
-      .replace(new RegExp(`#\\s+${escapeRegExp(title)}`), "")
-      .trimStart();
+    text = text.replace(/^.+(\n|$)/, "");
   }
 
   // Replace any <br> generated by the turndown plugin with escaped newlines
